NavigationPredictor: Use viewport instead of screen coordinates
We were recording the screenY value of the most recent pointer down,
and subtracting it from anchor element bounds in the root frame
coordinate space; leading to values that were incorrect. This CL
records the pointer down coords in the visual viewport's coord
space instead.

It also turns out that the anchor bounds we were using for local root
subframes was in the local root's coord space instead of the viewport
space. Existing methods in LocalFrameView don't support converting
this coords to the viewport space, so we now avoid doing
computations and recording metrics for such subframes (tests added
for this scenario as well).

Bug: 336852432
Change-Id: I8cb48b444c25f24ccdcd58428ce7d73cea6ce8e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5636365
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1317424}
NOKEYCHECK=True
GitOrigin-RevId: f872bd257c95fbb3d91b624bff9857ee7cad23f3