Add box-sizing: border-box for <option> in <select> dropdown
CL:5617111 set the minimum height of the <option> in <select> dropdown
to 24px.However, the default height exceeded 24px due to padding. This
CL adds box-sizing:border-box so that the padding is included in
min-block-size, which results in the default height actually being 24px.

Bug: 339141283
Change-Id: Ic5227c93f0d7be44c2b1bae572eb4e2bee5bb0c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5637883
Reviewed-by: Sanket Joshi <sajos@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Stephanie Zhang <stephanie.zhang@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1317432}
NOKEYCHECK=True
GitOrigin-RevId: 1566d974e40a7f28d59ac6813683d8c5054185da