Add DISPLAY_MEDIA_SYSTEM_AUDIO in content settings
This CL is aiming for building up a way to allow WebUI to access the
system audio by default. It enables better user experience for system
web apps on ChromeOS.

To do so, this CL adds a new ContentSettingsType
DISPLAY_MEDIA_SYSTEM_AUDIO so that when an app has this permission,
when calling getDisplayMedia() it can get the system audio
track without going through the display media selection picker.

Here are some consultation bugs we have for this proposal:
[Privacy]: b/337995113
[Security]: b/337996394

Bug: b:337770054
Test: Browser test

Change-Id: I0c7363712a3ca5decdc37e9b10eea5727361e7fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5519954
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1317775}
NOKEYCHECK=True
GitOrigin-RevId: da330e19b7d0290fd4a7d535e66ad8d8d39bd3ed