FLEDGE: Flag to enforce permission policy in contributeToHistogramOnEvent
This check was previously missing.

A histogram will always be collected to monitor how often this
is violated, and a devtools warning will always be produced.

When the flag is on, violating calls will be ignored, which is
a further compatibility compromise as they should throw.

Bug: 346766790

Change-Id: Iff8a74ead01f0bb52eddb7eb24eb727d5c33df56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5622811
Reviewed-by: Alex Turner <alexmt@chromium.org>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Orr Bernstein <orrb@google.com>
Cr-Commit-Position: refs/heads/main@{#1317915}
NOKEYCHECK=True
GitOrigin-RevId: 1b65e59609692e38fdb20d5d1369e951c2ab8869