Don't clamp fragmentation block-size if it actually fits.
The code that called SetIsBlockSizeForFragmentationClamped() incorrectly
assumed that we had (so far) decided not to move past the fragment, but
this wasn't checked. This would cause the block-size for fragmentation
to be clamped, setting an incorrect BlockSizeForFragmentation()
(LayoutResult), which in turn could result in false cache hits: a
fragment would appear to fit within the new and smaller fragmentatainer
block-size, although it didn't.

Bug: 348445899
Change-Id: Ia7b74ba4861c341d1236bafcf881d6745ce0c161
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5645005
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1317931}
NOKEYCHECK=True
GitOrigin-RevId: caa81df60fe607665b65b0c9653c051b8fde57a9