JNI Zero: Guard against accidentally returning ScopedJavaGlobalRefs
This can happen for Java->Native methods that are a part of a class,
since we do not control the method declaration.

We call .Release() on the returned value, so this ensures we are not
leaking global references.

Also fixes one spot where we were leaking global references.

Bug: 40944912
Change-Id: Ie009eb795ef6a3bafa3f09a2c976d9b6b831708e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5646786
Reviewed-by: Sam Maier <smaier@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Owners-Override: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1317980}
NOKEYCHECK=True
GitOrigin-RevId: 6241bbaec5aa99cce15f5c4a1164b58d31048873