Remove LayoutNG stuff from test runner.
We used to allow text differences if the only difference was LayoutNG
vs. non-LayoutNG in the object name. But the code here was no longer
correct. For one, the _is_render_tree failed to detect anything, because
size is no longer part of the text dumped for a layer, as of CL:4415405.
The _is_layer_tree check wasn't entirely correct, either, since not all
tests have indentation before "layers".

However, no tests depend on this anymore, so just remove it.

Bug: 40056555

Change-Id: I08e08cceb3f3027fd6f713e90670e30206cb0b18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5636436
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318030}
NOKEYCHECK=True
GitOrigin-RevId: 4313b9a7da1950d78bdab7d552438c6d83039a4f