[Blink] Pass internal usage through GetLastMailboxTextureAndSize()
WebGPUSwapBufferProvider::GetLastMailboxTextureAndSize() returns a new
wgpu::Texture associated with the Mailbox of the current (or last)
swapbuffer. This wgpu::Texture takes the TextureUsage passed in to the
WebGPUSwapBufferProvider on its construction. However, it currently has
no internal usage. This is problematic in its concrete use case [1],
where the texture thus created is passed to
CopyTextureToResourceProvider() [2] and must have the internal
usages that GpuCanvasContext originally determined would be necessary
for the swapbuffer texture in the context of that call.

This CL extends the flow in which the usage of
GetLastMailboxTextureAndSize() is set to work on the internal usage:

* WebGPUSwapBufferProvider takes in internal usage from the client
* GpuCanvasContext supplies the internal usage that it has computed
  will be necessary for its swapbuffer textures
* GetLastMailboxTextureAndSize() sets internal usage when creating a
  wgpu Texture

NOTE: I considered having WebGPUSwapBufferProvider take in the
TextureDescriptor, but GpuCanvasContext mutates it in the course of
execution [3].

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/webgpu/gpu_canvas_context.cc;l=149-157;drc=90cac1911508d3d682a67c97aa62483eb712f69a;bpv=1;bpt=1
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/webgpu/gpu_canvas_context.cc;l=884;drc=90cac1911508d3d682a67c97aa62483eb712f69a
[3] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/webgpu/gpu_canvas_context.cc;l=127;drc=90cac1911508d3d682a67c97aa62483eb712f69a;bpv=1;bpt=1

Bug: 339171225
Change-Id: Iadcff937049ff74fff73d80e69aa117fb3ebb1ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5644949
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318465}
NOKEYCHECK=True
GitOrigin-RevId: 23355f198d697b9ff8b6eb904d7ac006111f6eeb