Prevent SVGResourceDocumentContent load-completion racing with dispose
A call to Dispose() could be performed before the load had completed.
This meant that we had dropped the reference to the
SVGResourceDocumentContent and could not clean up the contained Page.

Set a flag in Dispose() and then use that bail with an error in
UpdateDocument() if it is set.

Bug: 345076820
Change-Id: I0eeedcad90bbe7a293e7d518ac3a23c96141ee18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5645897
Commit-Queue: Philip Rogers <pdr@chromium.org>
Auto-Submit: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318560}
NOKEYCHECK=True
GitOrigin-RevId: e314c045d35c1d0cd107fa57f0a9955c8ecc0e89