Move ReleaseLockedImages to CanvasResourceProvider::FlushCanvas
The main motivation for this change is to remove the call to
ReleaseLockedImages from Canvas2DLayerBridge::FlushRecording, since we
plan on deleting the whole Canvas2DLayerBridge class. The plan is for
flushes to be done by calling CanvasResourceProvider::FlushCanvas,
but since that function doesn't release locked images, we to add that
logic.

Currently, locked images aren't released consistently on all flushes
because not all calls to FlushCanvas have a corresponding call to
ReleaseLockedImages. This CL uniformises this by moving
ReleaseLockedImages to a centralized location that applies to all
flushes.

Bug: 40280152
Change-Id: Ice532733969b4895f05d44cf32e3a2226721ab8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5637530
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318628}
NOKEYCHECK=True
GitOrigin-RevId: 8070a21b442106bcbd8d9870d894b444a344d74f