Update branched-revision for WebXR samples
The webxr samples are a major fork of their upstream, the currently
listed revision (which does appear to be accurate) is from a now-deleted
branch and thus is not suitable for tracking the way that the revision
is intended to be used. This new commit is the base of the branch that
these were originally forked from and thus the more accurate and
appropriate commit to track, as it is from the `main` branch. There
is a later merge commit of the branch these were forked from, but that
would not be appropriate to include as it has other changes that
while they may have been essentially ported over, are not verifiable as
being included in our fork, and likely contains changes that are
definitely not in our fork.

Fixed: 348553821
Change-Id: Ie6607c64c803f50e492a069943d55fb503b09160
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5648627
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318630}
NOKEYCHECK=True
GitOrigin-RevId: 30471031ffc45fdcab53656b234abdae1bb6b91a