[css-anchor-position] Missing check for PositionAnchor
The presence of position-anchor will affect the passes in
OutOfFlowLayoutPart::CalculateOffset(), hence need to be taken into
account for storing old style to trigger transitions correctly.

This fixes the added test, but not the case in the bug report. Will need
to look further into that.

Bug: 344506496
Change-Id: I07938b325485f75de35b5f2fd382401257efe061
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5645058
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1318987}
NOKEYCHECK=True
GitOrigin-RevId: cb1fdee67ecc2951607d5a6a507718efe6c35067