[Blink] Set size on texture in GetLastMailboxTextureAndSize()
In the course of adding tests of this method, I discovered that we were
missing setting the size on the wgpu::Texture that it creates.

Note: Doing this enables simplifying the method to no longer return the
size separately. I'll do that in a followup.

Bug: 339171225
Change-Id: I7d44f4d7e885598f621aff6a44463cad920af170
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5648061
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319121}
NOKEYCHECK=True
GitOrigin-RevId: 1c83d7f939d63d3b9484db6c07f34b8b57e5156d