IWYU: returning an std::unique_ptr requires WebMediaPlayer type definition
Fixed missing include of WebMediaPlayer in WebLocalFrameClient.

But doing just that breaks root_scroller_test.cc in MacOS, that will
bring that dependency and include a macro fixed1 from Carbon. So
change the name of the variable to avoid the clash.

Bug: 41455655
Change-Id: I3b398f56783a4d273f54059b19fb7794c27b42db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5594110
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: José Dapena Paz <jdapena@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1319127}
NOKEYCHECK=True
GitOrigin-RevId: 278184684bf337675ca052b1fa86d4354ada0ef2