media: Return video region bounds for RequestVideoFrameAt()
This CL adds region bounds to be returned in RequestVideoFrameAt(),
which will be used for Lens overlay. This is needed to support the new
Google Lens search experience.

Also renamed the function to RequestVideoFrameAtWithBoundsDiagnostics()
for consistency with image search.

Manually tested with "Search Video Frame..." feature. With some
additional hacking, verified that the returned region is correct.
See b/344600237#comment7 for details.

Bug: b/344600237
Test: See above
Change-Id: I0ac227c80a9f871d9f06e3bb5678839da82d9d5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5641354
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319218}
NOKEYCHECK=True
GitOrigin-RevId: 52c71b21784508e8d859d91cfcb3bd641ce80cfe