Deprecate expectedImprovement from DelegatedInkTrailPresenter
This attribute was added so that we could potentially tell the developer
about the improvement they could potentially expect when using this API.
However, this is likely an avenue for finger printing, and we end up
asking ourselves "why wouldn't they want it?" So it should probably just
be removed. This is not expected to be a breaking change as
expectedImprovement was not actually implemented.

PR to update the spec: https://github.com/WICG/ink-enhancement/pull/34

Change-Id: I8aed5e56f8f3ed174013d25848dc52e01927ae7a
Bug: 347647410
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5627855
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Sahir Vellani <sahir.vellani@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1319557}
NOKEYCHECK=True
GitOrigin-RevId: 51aade06a8f25a1197bf7dd55921d760cebabe44