Remove the Python version of StringHasher.
Instead of computing these hashes in Python, we can make parts of
StringHasher constexpr. This means that we no longer need to maintain
two parallel versions of our hashing code, which will make it easier
to change it in the future. (Well, we still need a constexpr version of
the hashing code, but it seems easier to share logic between C++ and
constexpr C++.)

Change-Id: If3b97bdf9382d85a146253e04205805c2a9aff1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5658331
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319722}
NOKEYCHECK=True
GitOrigin-RevId: e2fedd42eb7602458829f501f62188942ec6471e