[Blink] Add SCANOUT usage for WebGPU SharedImages sent to display on Mac
I will be changing Blink to compute whether a TransferableResource is an
overlay candidate on Mac solely by whether SCANOUT SI usage is present
rather than by whether SCANOUT or WebGPU SI usage is present. In
preparation for making that change, this CL changes the place where
SharedImages are created that (a) have WebGPU usage and (b) may be sent
to the display compositor to also have SCANOUT usage.

This change is a no-op as far as the SharedImageBacking that will be
selected: SCANOUT usage would force IOSurfaceImageBacking to be chosen
on Mac, but WEBGPU usage already forces IOSurfaceImageBacking to be
chosen on Mac.

Bug: 348947993
Change-Id: I9c508407face8cdf206df49279acd164c69f64e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5650929
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319790}
NOKEYCHECK=True
GitOrigin-RevId: 2740f803e175f779b66ad0e4f84934777876f871