Add int8 quantization support to the denylist custom op
Updates the test to have the same repro case as in the bug, checking for
the same value that I get when running the model outside of chrome.

Note that I'm not incrementing MODEL_ENGINE_VERSION since this CL needs
to be merged to M127 and it would conflict with the 2.18 TFLite version
update that is in M128. This is safe since unlike the other two ops, the
denylist op does not return an error for unhandled data types.

Bug: 349440328
Change-Id: I13f43ca7cccdd531f612504ef4c50a8bcb195e7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5660277
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319842}
NOKEYCHECK=True
GitOrigin-RevId: 2bd964e7195743775ef408ca1b094f2fe8a7b126