Combine eligible source and trigger registrations in 1 redirect chain
This eliminates the need to enable debug reporting for the source and
wait on its corresponding source-success debug report to be emitted.

Note that we still tend to need the debug cookie to be present on the
source registration in order for the verbose-debug-report-under-test
for the trigger to be emitted.

Bug: 348594906
Change-Id: Iae37a84da94941940aa0abbafc39b24a12104288
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5660460
Reviewed-by: Nan Lin <linnan@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1319957}
NOKEYCHECK=True
GitOrigin-RevId: b5bc45aa34b5e0ca7bfda828f812a7392a17eb08