[RasterInducingScroll] Handle painting under effect escaping scroll
In the following case:
<div style="overflow: scroll">
  <div style="transform_under_scroll clip_under_scroll">
    <div style="opacity: 0.5">
      <div style="position: absolute; ...">not contained</div>
      <div>contained</div>
    </div>
  </div>
</div>
the effect applies to abspos, but the scroll doesn't.

This CL avoids switching to the effect's transform if the effect has
null OutputClip (which means something under the effect may escape the
current transform or clip). Then when we switch to the clip and the
transform, we'll detect the start of contained scrolling contents
correctly.

Bug: 40517276
Change-Id: I9d479065624eb890d6a6d77f35b33cb59a091c7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5643257
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320107}
NOKEYCHECK=True
GitOrigin-RevId: ed58811219653bf7a6e4a6ed56a21d11a41ce5f6