Add support for quotes in @page margin boxes content.
Since quotes create an extra level of layout objects, it will no longer
be a simple LayoutBlockFlow (page margin box) with leaf nodes, so that
the existing IsDetachedNonDomRoot() would fail when building the layout
object tree. Therefore change this flag (which is only used when DCHECKs
are enabled) to be set for the entire non-DOM tree, rather than just on
its root.

Bug: 40341678
Change-Id: Ie5e9ca34491b42dcff9ad029aa0fe8fedcce45ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5658437
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320273}
NOKEYCHECK=True
GitOrigin-RevId: a01c0a194a2afd96c9bcc9b9e51ca4d8f6cae9be