Update the canvas state and the FontDescription to match
After a font is set in Canvas, users can change the font properties,
such as FontKerning. One copy of the properties is saved in the
FontDescription and the another copy is saved in the
CanvasRenderingContext2DState. When font is updated, the copy in the
FontDescription get updated automatically, and the copy in the State
file stay unchanged. This inconsistency caused some properties
settings to fail.

Bug: 338965374

Change-Id: Ic2e72f8d1a5e5007648dfa7da7698d836ef54c42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5637432
Reviewed-by: Andres Ricardo Perez <andresrperez@chromium.org>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Commit-Queue: Yi Xu <yiyix@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320435}
NOKEYCHECK=True
GitOrigin-RevId: a510e56872fd8300dee9bc79d378cbfd271aec10