[SafeArea] Update safe-area-inset* for setter frame
DynamicSafeAreaInset is introduced in crrev.com/c/5597281, where the
Page will set a safe-area-inset* (SAI) to its main frame as the browser
controls scroll. However, this change does not cover cases where sub frame goes to fullscreen should also have their SAI applied.

Besides, when the SAI has an update from the browser, when DynamicSAI is
enabled, instead of always overriding the current SAI with the new one,
Page also needs to read the current browser controls state.

To fix the above issues, this CL passes the frame where
DisplayCutoutClientImpl that received the IPC is associated with the
Page. Then the Page will determine whether to apply the SAI directly to
the frame, or check the browser controls state before applying it.

This CL also adds check in Page::UpdateSafeAreaInsetWithBrowserControls
to ignore the call when a fullscreen element exists, as browser controls
should be gone when in fullscreen mode.

===

The are still issues regarding SAI isn't being reset properly when subframe enter / exit fullscreen mode; the fix is out of scope for this CL and will be addressed separately.

Bug: 324436581, 349187600
Change-Id: I28318c6adebee6568f97f0a1deea42f909787194
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5627700
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320539}
NOKEYCHECK=True
GitOrigin-RevId: 8b3db9ae37621fb6df0cc9bf2a0d13d8c16de535