Handle unexpected types in GetStringFromAttributeOrStringValue
In GetStringFromAttributeOrStringValue we should be able to assume
that `value` is either a CSSFunctionValue or a CSSStringValue, since
all other types of CSSValues produced in Content::ParseSingleValue
should have been handled by Content::ApplyValue before calling
GetStringFromAttributeOrStringValue. However, as observed
in the bug  there is some unexpected type that is not getting handled.

In this CL, update the code so that we don't crash in this case and
instead just produce an empty string for that part of the `content`
value.

Also, add two DCHECKs to try to help with trying to identify
the unexpected CSSValue.

The first DCHECK one tests the theory that the unexpected type is
coming from the ConsumeImage call in Content::ParseSingleValue,
which could cause a CSSLightDarkValuePair to be created if there
were a light-dark() function in a user-agent shadow DOM. I can't
find any user-agent shadow DOM where there is a light-dark()
inside a `content` rule, so this remains only a theory as to
the cause of the crash.

The second DCHECK will hit if this first theory is wrong and
`value` has some other unexpected type.

Bug: 348304397
Change-Id: I9d5e4ca8483a48d171ddc02c20becc95bf698e2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5661598
Commit-Queue: Dan Clark <daniec@microsoft.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320691}
NOKEYCHECK=True
GitOrigin-RevId: ad5a1ca92676cd00eeb366e7cc9fd62ceed62ed8