Add UMAs for data URL length and parse time
This CL adds a variant of the existing Blink.Network.ParseDataURLTime.*
histograms for image resources.

This also adds a Blink.Network.DataUrlLength.* histogram which measures
the length of data URLs, so we can see a distribution of the size of
data URLs we parse from the field.

This also moves the logic to record the metrics to the end of the
function, rather than on cleanup as we don't want to measure metrics
when we get net error.

Bug: 348442535
Change-Id: Iac009151185f77a10d1d15797f3939db238d2049
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5660201
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Nidhi Jaju <nidhijaju@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320725}
NOKEYCHECK=True
GitOrigin-RevId: 817e12e8ff7084f815c12f7c9b43b306e0d25e93