[Blink] Test WebGPU swapbuffer passes internal usage in GetNewTexture()
This CL tests that WebGPUSwapbufferProvider passes the internal usage
with which it was created as well as RenderAttachment usage to support
lazy clearing when associating a mailbox in GetNewTexture().

Note: This CL updates and extends an existing test, which I realized
wasn't actually testing the use case I had intended it to. I'll add
a true test of that use case next.

Bug: 339171225
Change-Id: I2ff25d23d2ebbf64e32a3889b8a4ab7ccae0b469
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5666495
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320850}
NOKEYCHECK=True
GitOrigin-RevId: 6e22876b0d411f73aee0120e909856b58873d1db