[Blink] Clear CanvasResourceRasterSI::texture_target_ on SI destruction
CanvasResourceRasterSharedImage::texture_target_ caches the texture
target of the resource's ClientSI. This CL clears it when that SI is
destroyed, as at that point it no longer has semantic meaning. This
paves the way for a followup CL that eliminates this field.

Bug: 41494843
Change-Id: Id66a7e47e0db75079b79a3fe8023f241a876dcc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5664716
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320917}
NOKEYCHECK=True
GitOrigin-RevId: 38412ad20aa69ac1284934b95de2cb36a7813d7b