Fenced frame: attempt to deflake ARA network revocation test.
A recently introduced "revoke-attribution-reporting.sub.https.html" test
file is causing multiple flakes, triggering multiple bugs. This CL
attempts to de-flake the test simply by moving it to a different
directory. Reasoning below:

The try bots generally run WPTs on multiple (8 to be exact) workers, so
there are 8 WPTs running at any given time. On top of that, it tends to
put tests in the same directory on the same worker so that they run in
series. The test was originally in the "wpt_internal/fenced_frame"
directory, so it would run in series with the other fenced frame WPT
internal tests.

ARA is set up to go to a "/.well-known/" endpoint. This endpoint is
shared across all ARA WPTs, both internal and external. However, the
tests are set up to expect only 1 test to be interacting with the
endpoint at any given time. The "attribution_reporting_promise_test"
test function wipes any existing reports in the endpoint's stash before
running the test, and pollEventLevelReports()/pollAggregatableReports()
(as well as handle_get_reports in
"attribution-reporting/resources/reports.py") assumes that only the
reports from that one specific test will be in place.

This hasn't been an issue before, since all the ARA-related tests are in
one directory and run in series. But the introduction of an ARA test to
the fenced frame test directory has started causing issues, since it
resulted in tests running in parallel and causing collisions with the
ARA endpoint. This resulted in things like the report being cleared by a
different test before having a chance to check for it, or a report sent
from a different test making it into the current test.

The cleanest (but most time consuming) way to fix this would be to
refactor the ARA endpoints to be able to handle concurrent tests
interacting with it. But since this is causing a lot of build failures
and needs to be fixed quickly, this test opts to instead move the
problematic test into the ARA WPT directory so that it runs in series
with the rest of the tests that talk to the ARA endpoints.

I believe this same test is what triggered this bug to be filed:
https://crbug.com/349397580

The effort to do the full refactor is being tracked with this bug:
https://crbug.com/350084227

Bug: 348613363
Bug: 349265566
Change-Id: I27d936e207a1da7fd0535fd67254eee659177491
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5666818
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Liam Brady <lbrady@google.com>
Cr-Commit-Position: refs/heads/main@{#1321228}
NOKEYCHECK=True
GitOrigin-RevId: 0bac98b4db06e32f598150bda4e6020c1f6bf898