[PEPC] Disable PEPC if recalculating style causes layout change
This CL complements the case in which the PEPC styling is changed,
causing a move that but then move back instantaneously in
requestAnimationFrame callback.
Based on https://github.com/w3c/IntersectionObserver/issues/263, it
seems like if the re-layout happen in the requestAnimationFrame
callback, IntersectionObserver won't send us a notification about a
visibility change event, even if the layout totally hides the PEPC.

We had another measure, calculating the intersection rect with the
viewport using lifecycle update events, but that didn't work either
because the intersection rect with the viewport stayed the same.
So, we're making sure that even styling changes that affect the
intersection rect will trigger the cooldown time, no matter if they
happen in requestAnimationFrame or not.

Fixed: 347509736, 348359040
Change-Id: I98dde8d42d696912d499b83be309630a1f4c9392
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5658195
Reviewed-by: Yi Gu <yigu@chromium.org>
Commit-Queue: Thomas Nguyen <tungnh@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1322092}
NOKEYCHECK=True
GitOrigin-RevId: b45d131eeb3a693b1234898b3b6c48c304d73dc3