Implement getActualBoundingBox feature for TextMetrics
This change creates a new API method for TextMetrics objects in canvas
that uses font data and precalculated text runs, to calculate the
bounding box for an interval of characters. This interval is defined
by character positions in the input text.

The method returns a single rectangle, the actual bounding box of the
substring of the input text defined by the input indexes.

The new API was enabled under the `ExtendedTextMetrics` flag.

Bug: 341213359
Change-Id: I289a52283d0bd5a0001c5ec7a514cf16abba9611
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5546024
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Andres Ricardo Perez <andresrperez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1322331}
NOKEYCHECK=True
GitOrigin-RevId: a19e45fae5851306ea33f38abbb5853eb6783de1