[CompositeClipPathAnimation] Immediately invalidate paint
This is a follow up to "Make Composited Animations Invalidate Paint Directly." While that version of the change made crashes less likely to occur, crashes still appear in clients with the experiment with the same stack trace caused by the paint properties of a composited animation being invalidated.

This change removes the process of delaying invalidation for composited animations entirely and simply invalidates paint immediately when an animation is canceled or set compositor pending.

Bug: 340928289

Change-Id: I33c74d341e74549aa6d5a96e815c0705f03c2a72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5583870
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Claire Chambers <clchambers@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1322351}
NOKEYCHECK=True
GitOrigin-RevId: 60514ef1b63df5fa7aea3af87f6797ab9dcb5e32