Let IsSyncTree() work properly in unit tests
Previously many unit tests didn't set commit_to_active_tree to false,
but created pending tree which failed the IsSyncTree() check because
the active tree instead of the pending tree is the sync tree when
commit_to_active_tree is true. These tests are actually testing the
commit-to-pending-tree configuration. Due to the failure, we couldn't
CHECK(IsSyncTree()) when we are committing in
PictureLayerImpl::UpdateRasterSource.

- Let commit_to_active_tree be false by default for most unit tests,
  and set it to true for unit tests requiring that.
- Let most LayerTreeHostImpl unit tests be parameterized to test both
  commit_to_active_tree and commit_to_pending_tree configurations.

Change-Id: Ide8c2a92031579e68231574eb273b8c1a853a07a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5671632
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1322492}
NOKEYCHECK=True
GitOrigin-RevId: 4b76f73be60dae897cbab8102dd34450dfa05cc1