Update UMA and Perfetto Traces for partitioned :visited links
This CL adds the following metrics:
(1) Counterparts to UMA that already exist in VisitedLinkWriter, adding them to PartitionedVisitedLinkWriter so they can be compared during the Finch experiment.
(2) New UMA to track :visited-ness and the number/code-path of per-origin salts.
(3) Perfetto Traces on PartitionedVisitedLinkWriter init and add functions as can be found in VisitedLinkWriter.

The goal of all these additions is to prepare for a Finch experiment.

Change-Id: I1b4a05ee6b2fbdc02bf40574670252055485186f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5645756
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Kyra Seevers <kyraseevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1322732}
NOKEYCHECK=True
GitOrigin-RevId: 3231ed7a06ffbb6c1c75cf5346c7334918140c87