Replace variable-length arrays with hard-coded lengths
Various functions create variable-length arrays whose length is based on
the render quantum size. Currently, this size is hardcoded as 128;
however, in the future it will be user-definable (crbug.com/40637820).

We remove the VLAs by setting their length to the hardcoded value. In
order to be future-proof, we additionally add runtime checks to ensure
that this length is correct. The original expression determining the length of the arrays is the first argument to the CHECK_EQ statement, allowing this commit to be easily undone when the time comes.

Bug: 349656479, 40637820
Change-Id: I5dda6b1a46f664733d5bf0f1b5fa68244abf06e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5666109
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Michael Wilson <mjwilson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1322746}
NOKEYCHECK=True
GitOrigin-RevId: 2f3bc4660da815a6753cbd97fb24168db4cf9b24