Remove font reliance on incremental StringHasher.
We are planning to get rid of the AddCharacter* interfaces
in StringHasher (incremental hashing), so remove the use of them
from the font code.

Longer-term, it would also be nice to get rid of the
WTF::Add*ToHash() code (it uses djb2, which isn't designed
for this at all), but that's a slightly bigger cleanup.

Change-Id: Iece6ab751156befb5a6bfcff0c673eb3ae1fb04c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5674543
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1322846}
NOKEYCHECK=True
GitOrigin-RevId: 19de47d979af56a5d5c656d8be346ea22ec4a2e9