Remove unused performance side panel
We had initially planned to implement features here earlier this year
but plans changed and it is TBD when or if we will use a side panel.
Deleting this code to reduce dead code in the codebase and reduce
maintenance costs.

Bug: 346684380,346694145
Change-Id: If0c24dfd7c88ebfebcac500ec3c923a3c88ee7db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5670606
Commit-Queue: Alison Gale <agale@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: John Lee <johntlee@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1322966}
NOKEYCHECK=True
GitOrigin-RevId: 9c8beafc9680a9aee1eaf32978781a88f80fa828