Remove obsolete test
composited-animations-translate-rotate-scale is flaking on the Mac.
The test is not particularly relevant since it uses the internals.pause
API, which is not used in production code. Paused animations do not
run on the compositor in practice.

We have equivalent test coverage in WPT in css/css-transforms/animation.

See transform-interpolation-(translate|scale|rotate|skew|...)

Bug: 350496178
Change-Id: I30dde31638d5b6f4f2105b0903b6fe655f3a71a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5678776
Reviewed-by: David Awogbemila <awogbemila@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1323415}
NOKEYCHECK=True
GitOrigin-RevId: 1b62b71affb6c3940fafc4d918f6bb042ae3a158